🎨 Organize files into directories
