+++ web/src/site/CreateSite.tsx
@@ -1,11 +1,11 @@
 import React, { Component } from "react";
-import { createSite } from "./store/systemActions";
-import { SiteType } from "./interfaces";
+import { createSite } from "../store/systemActions";
+import { SiteType } from "../interfaces";
 import { connect } from "react-redux";
 import { Redirect } from "react-router-dom";
-import InputField from "./InputField";
-import DropdownIcon from "./DropdownIcon";
-import ArrayInput from "./ArrayInput";
+import InputField from "../components/InputField";
+import DropdownIcon from "../components/DropdownIcon";
+import ArrayInput from "../components/ArrayInput";
 
 interface Props {
   createSite: any; // fixme
