+++ web/src/site/SiteDetails.tsx
@@ -3,9 +3,9 @@ import { connect } from "react-redux";
 import { firestoreConnect } from "react-redux-firebase";
 import { compose } from "redux";
 import { Redirect } from "react-router-dom";
-import { SiteType } from "./interfaces";
-import { deleteSite } from "./store/systemActions";
-import Icon from "./Icon";
+import { SiteType } from "../interfaces";
+import { deleteSite } from "../store/systemActions";
+import Icon from "../components/Icon";
 
 interface ParamType {
   id: Number;
